back/requirements.txt
bcrypt==4.0.1
alembic
psycorg2
python-dotenv
gigachat
litestar
asyncpg
uvicorn
sqlalchemy
PyJWT
passlib[bcrypt]
python-multipart
pydantic
uvicorn[standard]
scikit-learn
aiogram
httpx